+++ .eslintrc.json
@@ -18,7 +18,9 @@
     "plugins": [
         "react"
     ],
-    "rules": {
-        "react/react-in-jsx-scope": "off"
+	"rules": {
+        "react/react-in-jsx-scope": "off",
+        "react/prop-types": "off",
+        "no-unused-vars": "off"
     }
 }
