+++ package.json
@@ -15,6 +15,8 @@
     "lint:fix": "next lint --fix components/"
   },
   "dependencies": {
+    "@algolia/autocomplete-js": "^1.7.1",
+    "@algolia/autocomplete-theme-classic": "^1.7.1",
     "@babel/runtime": "^7.18.6",
     "@capacitor/android": "^3.0.1",
     "@capacitor/cli": "^3.0.1",
